Add support for required version.
Check the appcastItem's enclosure for a field labeled
"sparkle:requiredVersion". If found, compare that version to the host's
currently-running version. If the host is not running at least the
required version, do not allow it to launch without an update.  This
means hiding the "Skip" and "Later" buttons in the Update alert, and
terminating the program if the user closes the update alert dialog (via
the window control) without launching the update.